# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel News API",
        "description": "Документація для тестового завдання REST API новин",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Локальний сервер"
        }
    ],
    "paths": {
...
        "/api/news": {
            "get": {
                "tags": [
                    "Public News"
                ],
                "summary": "Отримати список усіх опублікованих новин",
                "description": "Дозволяє шукати новини за текстом та фільтрувати за автором.",
                "operationId": "c33e145d81ca885577167b753733eca0",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Пошук за заголовком або коротким описом",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "author_id",
                        "in": "query",
                        "description": "ID автора (користувача)",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успішна відповідь із пагінацією"
                    }
                }
            }
        },
        "/api/news/{id}": {
            "get": {
                "tags": [
                    "Public News"
                ],
                "summary": "Переглянути конкретну публічну новину",
                "operationId": "279bf6ef5af47ef92aee699942813c0b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID новини",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успішно"
                    },
                    "404": {
                        "description": "Новину не знайдено або вона ще не опублікована"
                    }
                }
            }
        },
        "/api/my-news": {
            "get": {
                "tags": [
                    "My News"
                ],
                "summary": "Отримати список власних новин",
                "description": "Повертає пагінований список усіх новин (включаючи чернетки), створених поточним користувачем.",
                "operationId": "70f1ac503c7d466c1e93155a7f89c789",
                "responses": {
                    "200": {
                        "description": "Успішно"
                    },
                    "401": {
                        "description": "Неавторизовано"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "My News"
                ],
                "summary": "Створити нову новину",
                "description": "Створення новини з мультимедійними блоками через multipart/form-data.",
                "operationId": "70865acb358bb80f3a70cf928d6b357b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "short_description",
                                    "is_published"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Нова новина"
                                    },
                                    "short_description": {
                                        "type": "string",
                                        "example": "Короткий опис"
                                    },
                                    "is_published": {
                                        "type": "boolean",
                                        "example": 1
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "blocks[0][type]": {
                                        "type": "string",
                                        "example": "text"
                                    },
                                    "blocks[0][text_content]": {
                                        "type": "string",
                                        "example": "Зміст блоку"
                                    },
                                    "blocks[1][type]": {
                                        "type": "string",
                                        "example": "image"
                                    },
                                    "blocks[1][image]": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Створено"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/my-news/{id}": {
            "get": {
                "tags": [
                    "My News"
                ],
                "summary": "Переглянути свою новину",
                "operationId": "04a98d08bb6b1af0f348aad9e7449997",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успішно"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "My News"
                ],
                "summary": "Оновити існуючу новину",
                "description": "Для оновлення з файлами використовуйте POST та поле _method=PUT.",
                "operationId": "df6a4bc6827db0c9f322fdf6343eb34b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "_method": {
                                        "type": "string",
                                        "example": "PUT"
                                    },
                                    "title": {
                                        "type": "string"
                                    },
                                    "blocks[0][type]": {
                                        "type": "string",
                                        "example": "text"
                                    },
                                    "blocks[0][text_content]": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Оновлено"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "My News"
                ],
                "summary": "Видалити новину",
                "operationId": "b29fc32d17bcd5c121f99f630fb7a048",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Видалено"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Введіть токен, отриманий при логіні",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}